feat:MNST-10 refactored the variable volumeNumber

+++ src/main/java/com/spring/boot/project/ms/manga/store/application/dto/request/VolumeRequestDto.java
@@ -17,7 +17,7 @@ public record VolumeRequestDto(
 
         @NotNull(message = "the volume number is required")
         @PositiveOrZero(message = "the volume number must be 0 or greater")
-        Float volumeNumber,
+        Double volumeNumber,
 
         @NotBlank(message = "the title can't be empty")
         String title,
